为MOE和GMM 调整attr



# message auto-generated for no-merge-commit merge:
!2724 merge fix_get_attr_int_moe_gmm into master

为MOE和GMM 调整attr

Created-by: weixin_44156099
Commit-by: weixin_44156099
Merged-by: cann-robot
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->
grouped_matmul，本地ATK验证(GroupedMatmul_A4W4)：
![image.png](https://raw.gitcode.com/user-images/assets/7673863/89732078-633f-422d-a308-020285ac60a5/image.png 'image.png')
moe_compute_expert_tokens，dev仓ATK用例，：
![image.png](https://raw.gitcode.com/user-images/assets/7673863/a7ab6ea4-fdab-418f-a94b-19c6678c08a5/image.png 'image.png')
moe_token_permute_with_routing_map_grad，本地ATK验证：
![image.png](https://raw.gitcode.com/user-images/assets/7673863/e20fbccd-1a8c-4b29-8275-14d69cd2b79c/image.png 'image.png')
本地UT测试
![image.png](https://raw.gitcode.com/user-images/assets/7673863/d626e668-6594-4d17-80d3-01256d497873/image.png 'image.png')
## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [ ] ✨ 新特性
- [ ] ⚡ 性能优化
- [ ] ♻️ 重构
- [ ] 🧪 测试
- [ ] 📦 构建/CI
- [ ] 🔧 配置变更
- [ ] 📝 文档更新
- [ ] ⬆️ 依赖升级
- [x] 🔒 安全修复
- [ ] 🧹 代码清理
- [ ] ❓ 其他，请描述：



See merge request: cann/ops-transformer!2724

+++ gmm/grouped_matmul/op_host/op_tiling/grouped_matmul_tiling.h
@@ -195,7 +195,7 @@ private:
     bool isSingleY_ = false;
     bool isAllSingleTensor_ = false;
     bool hasBias_ = false;
-    int32_t groupType_ = 0;
+    int64_t groupType_ = 0;
     int64_t splitItem_ = 0L;
     uint32_t groupListType_ = 0;
     uint32_t xKDim_ = 0;
